customize argument to retrieve version
Programs differ in their options to output the version. Though  one of
-v, -V and --version should be appropriate for the vast majority of
programs, automatically determining the correct parameter is
extraordinarily difficult.

Trying to execute the program with the switch may lead to possibly
severe unintended side effects or output any unwanted output that
incidentally looks like a version string.

Parsing the help text output is also not likely to yield reliable
results. The formatting of the help text output is entirely up to the
author of the program and sometimes entirely missing. On top of that,
the same problem repeats itself as some programs use -h, some --help and
others expect to be called without any parameter to output help.

So the last remaining option feasible for a first implementation is
leaving the choice to the user.