customize version pattern
Some programs hide their version string in a whole bunch of output,
which may confuse the default pattern. While it may be too complicated
to leave the entire parsing of this output to the user, leaving the
conversion to the version function leaves customizing the pattern. This
already goes great lengths even when only the entire matching group is
considered, as look-around expressions can pin this group into a
context, enabling clearly filtering out the version string even from
complicated outputs.