recommend automated version determination
The find_version function is now mature enough to be the recommended way
of finding versions. Statically defining them is not very relevant to
the real use case of this tool.

The more complicated case of specifying an explicit versions dictionary
to the task generator is still documented further below and shown in the
example program.

The program used as an example was changed from ssh to bash because bash
understands --version, which is the default, while for ssh an argument
would be necessary to change this to -V.