provide extension rule
This allows more comfortable specification of version processing,
provided aptly named files, and especially avoids the unusual direct
call to create_task from out of the build function.

For complicated situations, it may be worthwhile to implement a feature
in the future, similar to the "subst" feature from waflib.TaskGen.